Remove sync_device from update loop
Remove sync_device API call from update loop entirely, to not hammer Meural API with an expensive call. Current item data is retrieved via get_item based on local current item id. Item displayed in the Meural web-interface will now become out of sync with device again - but this is an issue for Meural to solve, not within scope of this integration.