Improve play_media
Improve play_media by using device_load_item to load the item via the Meural API if the item is not in the current playlist.

Set _pause_duration when device is added, for edge case where media_play is called while _pause_duration is still initialized to 0.

Update README.md and add images.